Extraction of needed helpers to GearControllerHelper
