+++ app/helpers/google_analytics_helper.rb
@@ -1,0 +1,10 @@
+module GoogleAnalyticsHelper
+  def insert_google_analytics(track, public_view = false, custom_vars = {})
+    if !Cartodb.config[:google_analytics].blank? && !Cartodb.config[:google_analytics][track].blank? && !Cartodb.config[:google_analytics]["domain"].blank?
+      ua = Cartodb.config[:google_analytics][track]
+      domain = Cartodb.config[:google_analytics]["domain"]
+
+      render(:partial => 'shared/analytics', :locals => { ua: ua, domain: domain, custom_vars: custom_vars, public_view: public_view })
+    end
+  end
+end
