+++ gears/carto_gears_api/lib/carto_gears_api/gear_controller_helper.rb
@@ -1,0 +1,11 @@
+# This should be included in every controller using the following CARTO layouts:
+# - application
+module CartoGearsApi
+  module GearControllerHelper
+    include SafeJsObject
+    include CartoDB::ConfigUtils
+    include TrackjsHelper
+    include GoogleAnalyticsHelper
+    include HubspotHelper
+  end
+end
